+++ packages/forms/src/components/FlightStatsInput/debounce.js
@@ -1,0 +1,9 @@
+export default function debounce(callback, time = 500) {
+    let timer;
+    return function debounced(...args) {
+        clearTimeout(timer);
+        timer = setTimeout(() => {
+            callback.call(this, ...args);
+        }, time);
+    };
+}