+++ packages/listview/src/components/ColoredStatusColumn/index.js
@@ -1,8 +1,7 @@
 import React, { useCallback, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
-import InternalOverlay from 'react-rainbow-components/components/InternalOverlay';
-import { RenderIf } from 'react-rainbow-components';
-import { useOutsideClick, useWindowResize } from 'react-rainbow-components/libs/hooks';
+import { InternalOverlay, RenderIf } from 'react-rainbow-components';
+import { useOutsideClick, useWindowResize } from '@rainbow-modules/hooks';
 import { HeaderOption } from './option';
 import getBackgroundColor from './helpers/getBackgroundColor';
 import getColor from './helpers/getColor';